requirements.txt
google-generativeai
pyyaml
tqdm